backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
psycopg2-binary==2.9.7
python-decouple==3.8
google-generativeai==0.3.2
django-filter==23.3
djangorestframework-simplejwt==5.3.0
python-dotenv==1.0.0
celery==5.3.4
redis==5.0.1
nltk==3.8.1
textblob==0.17.1
scikit-learn==1.3.1
pandas==2.1.3
numpy==1.25.2
python-dateutil==2.8.2
requests==2.31.0
gunicorn==21.2.0
whitenoise==6.6.0
